+++ tests/test_annotation_tilerendering.py
@@ -2,21 +2,73 @@
 AnnotationRenderer and AnnotationTileGenerator
 """
 from pathlib import Path
+from typing import List, Union
 
 import matplotlib.cm as cm
 import matplotlib.pyplot as plt
 import numpy as np
 import pytest
 from PIL import Image
-from scipy.ndimage.measurements import label
-from shapely.geometry import MultiPoint
+from scipy.ndimage import label
+from shapely.geometry import LineString, MultiPoint, Polygon
+from shapely.geometry.point import Point
 from skimage import data
 
-from tiatoolbox.annotation.storage import Annotation, SQLiteStore
+from tests.test_annotation_stores import cell_polygon
+from tiatoolbox.annotation.storage import Annotation, AnnotationStore, SQLiteStore
 from tiatoolbox.tools.pyramid import AnnotationTileGenerator
 from tiatoolbox.utils.env_detection import running_on_travis
 from tiatoolbox.utils.visualization import AnnotationRenderer
 from tiatoolbox.wsicore import wsireader
+
+
+@pytest.fixture(scope="session")
+def cell_grid() -> List[Polygon]:
+    """Generate a grid of fake cell boundary polygon annotations."""
+    np.random.seed(0)
+    return [
+        cell_polygon(((i + 0.5) * 100, (j + 0.5) * 100)) for i, j in np.ndindex(5, 5)
+    ]
+
+
+@pytest.fixture(scope="session")
+def points_grid(spacing=60) -> List[Point]:
+    """Generate a grid of fake point annotations."""
+    np.random.seed(0)
+    return [Point((600 + i * spacing, 600 + j * spacing)) for i, j in np.ndindex(7, 7)]
+
+
+@pytest.fixture(scope="session")
+def fill_store(cell_grid, points_grid):
+    """Factory fixture to fill stores with test data."""
+
+    def _fill_store(
+        store_class: AnnotationStore,
+        path: Union[str, Path],
+    ):
+        """Fills store with random variety of annotations."""
+        store = store_class(path)
+
+        cells = [
+            Annotation(cell, {"type": "cell", "prob": np.random.rand(1)[0]})
+            for cell in cell_grid
+        ]
+        points = [
+            Annotation(point, {"type": "pt", "prob": np.random.rand(1)[0]})
+            for point in points_grid
+        ]
+        lines = [
+            Annotation(
+                LineString(((x, x + 500) for x in range(100, 400, 10))),
+                {"type": "line", "prob": 0.75},
+            )
+        ]
+
+        annotations = cells + points + lines
+        keys = store.append_many(annotations)
+        return keys, store
+
+    return _fill_store
 
 
 def test_tile_generator_len(fill_store, tmp_path):
